requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
et-xmlfile==1.1.0
Faker==8.10.0
idna==2.10
importlib-metadata==4.6.1
more-itertools==8.8.0
openpyxl==3.0.7
pluggy==0.13.1
py==1.10.0
pytest==4.5.0
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
zipp==3.5.0
